BUG: fix epsilon decay & pixel colours + docker improvements (#8)
Closes #3 

Docker improvements & specific Config.wtf file seemed to have fixed incorrect RGB pixel colours.

Epsilon decay was not implemented by default in the DQN model so added that. The `collect_step` function was also not using the correct policy so set it to use random actions in initial collect & the DQN policy otherwise.